Refactor parser module into cohesive units
- Extracted `src/parser/expressions.rs` for expression parsing
- Extracted `src/parser/statements.rs` for regular statement parsing
- Extracted `src/parser/declarations.rs` for type/trait/test declarations
- Extracted `src/parser/common.rs` for shared error types and helpers
- Updated `src/parser/mod.rs` to act as a facade and dispatcher
- Improved module encapsulation by using `pub(crate)` visibility

This change addresses the "Blob" structural smell in the parser module, making the code more maintainable and easier to navigate.
